pysdm constants test
