+++ test/Atmos/Parameterizations/Microphysics/KM_PySDMachine/README.md
@@ -1,0 +1,13 @@
+### TODOs:
+
+# PySDM.rhod != CliMa.rho
+
+# add 2 Displacements: 1 for Condensation and 1 for Advection
+# upd ClimateMachine state vars:
+     passing effective_radius to ClimateMachine (Auxiliary)
+
+
+
+# instead of liquid_water_mixing_ratio should be liquid_water_specific_humidity
+# AJ shouldnt we compute new e_int and new T based on new pp?
+
