+++ test/Atmos/Parameterizations/Microphysics/KM_PySDMachine/test/utils/KM_CliMa_no_saturation_adjustment.jl
@@ -1,12 +1,15 @@
 
-include("../KinematicModel.jl")
-include("./PySDMCall/PySDMCallback.jl")
-include("./KM_PySDM/KM_PySDM.jl")
+include("../../../KinematicModel.jl")
+include("../PySDMCall/PySDMCallback.jl")
 
 using .PySDMCallbacks
 using .PySDMCallbacks.PySDMCall
 
 using StaticArrays
+using Statistics
+using Test
+
+__init__()
 
 function vars_state(m::KinematicModel, ::Prognostic, FT)
     @vars begin
@@ -105,27 +108,16 @@ function nodal_update_auxiliary_state!(
         aux.e_pot = _grav * aux.z_coord
         aux.e_int = aux.e_tot - aux.e_kin - aux.e_pot
 
-        # saturation adjustment happens here
 
         # no saturation adjustment
-
-        #ts = PhaseEquil(param_set, aux.e_int, state.ρ, aux.q_tot)
-        #q = PhasePartition(ts)
         q = PhasePartition(aux.q_tot, 0.0, 0.0)
         aux.T = air_temperature(param_set, aux.e_int, q)
 
-        #aux.T = ts.T
-
         aux.theta_liq_ice = liquid_ice_pottemp(param_set, aux.T, state.ρ, q) #ts.T, state.ρ, q)
         aux.theta_dry = dry_pottemp(param_set, aux.T, state.ρ) #ts.T, state.ρ)
-        aux.q_vap = vapor_specific_humidity(q) # zmienne w przestrzeni
+        aux.q_vap = vapor_specific_humidity(q) #changes in space
         aux.q_liq = q.liq
         aux.q_ice = q.ice
-
-        # TODO: add super_saturation method in moist thermo
-        #aux.S = max(0, aux.q_vap / q_vap_saturation(ts) - FT(1)) * FT(100)
-        #aux.S_liq = max(0, supersaturation(ts, Liquid()))
-        #aux.RH = relative_humidity(ts)
     end
 end
 
@@ -182,34 +174,35 @@ end
 
 source!(::KinematicModel, _...) = nothing
 
-function main()
+function set_up_machine(
+    domain_extents_3d,
+    domain_resolution_3d,
+    t_end,
+    dt,
+    qt_0,
+)
     # Working precision
     FT = Float64
     # DG polynomial order
     N = 4 # 1 2 regular cells
     # Domain resolution and size
-    Δx = FT(20)
-    Δy = FT(1)
-    Δz = FT(20)
+    Δx = domain_resolution_3d[1]
+    Δy = domain_resolution_3d[2]
+    Δz = domain_resolution_3d[3]
     resolution = (Δx, Δy, Δz)
     # Domain extents
-    xmax = 1500
-    ymax = 10
-    zmax = 1500
+    xmax = domain_extents_3d[1]
+    ymax = domain_extents_3d[2]
+    zmax = domain_extents_3d[3]
     # initial configuration
     wmax = FT(0.6)  # max velocity of the eddy  [m/s]
     θ_0 = FT(289) # init. theta value (const) [K]
     p_0 = FT(101500) # surface pressure [Pa]
     p_1000 = FT(100000) # reference pressure in theta definition [Pa]
-    qt_0 = FT(7.5 * 1e-3) # init. total water specific humidity (const) [kg/kg]
     z_0 = FT(0) # surface height
 
     # time stepping
     t_ini = FT(0)
-    t_end = FT(60 * 30)
-    dt = 10 # was 40
-    output_freq = 9
-    interval = "90steps"
 
     # periodicity and boundary numbers
     periodicity_x = true
@@ -255,146 +248,5 @@ function main()
         #Courant_number = CFL,
     )
 
-    mpicomm = MPI.COMM_WORLD
-
-    # output for paraview
-    # initialize base prefix directory from rank 0
-    vtkdir = abspath(joinpath(ClimateMachine.Settings.output_dir, "vtk"))
-    if MPI.Comm_rank(mpicomm) == 0
-        mkpath(vtkdir)
-    end
-    MPI.Barrier(mpicomm)
-
-    model = driver_config.bl
-    vtkstep = [0]
-    cbvtk = GenericCallbacks.EveryXSimulationSteps(output_freq) do
-        out_dirname = @sprintf(
-            "new_ex_1_mpirank%04d_step%04d",
-            MPI.Comm_rank(mpicomm),
-            vtkstep[1]
-        )
-        out_path_prefix = joinpath(vtkdir, out_dirname)
-        @info "doing VTK output" out_path_prefix
-        writevtk(
-            out_path_prefix,
-            solver_config.Q,
-            solver_config.dg,
-            flattenednames(vars_state(model, Prognostic(), FT)),
-            solver_config.dg.state_auxiliary,
-            flattenednames(vars_state(model, Auxiliary(), FT)),
-        )
-        vtkstep[1] += 1
-        nothing
-    end
-
-    # output for netcdf
-    boundaries = [
-        FT(0) FT(0) FT(0)
-        xmax ymax zmax
-    ]
-    interpol = ClimateMachine.InterpolationConfiguration(
-        driver_config,
-        boundaries,
-        resolution,
-    )
-    dgngrps = [
-        setup_dump_state_diagnostics(
-            AtmosLESConfigType(),
-            interval,
-            driver_config.name,
-            interpol = interpol,
-        ),
-        setup_dump_aux_diagnostics(
-            AtmosLESConfigType(),
-            interval,
-            driver_config.name,
-            interpol = interpol,
-        ),
-    ]
-    dgn_config = ClimateMachine.DiagnosticsConfiguration(dgngrps)
-
-
-    # configuring PySDM
-    krnl = PySDMKernels()
-
-    spectra = PySDMSpectra()
-
-    rho_STP = 1.2252141358659048
-    micrometre = 1e-6
-    centimetre = 0.01
-    spectrum_per_mass_of_dry_air = spectra.Lognormal(
-        norm_factor = 60 / centimetre^3 / rho_STP,
-        m_mode = 0.04 * micrometre,
-        s_geom = 1.4,
-    )
-
-    n_sd = 25
-
-    pysdmconf = PySDMConfig(
-        (xmax, zmax),
-        (Δx, Δz),
-        t_end,
-        solver_config.dt,
-        n_sd,
-        1,
-        krnl.Geometric(collection_efficiency = 1),
-        spectrum_per_mass_of_dry_air,
-    )
-
-    pysdm_cw = PySDMCallWrapper(pysdmconf, init!, do_step!, nothing)
-
-    pysdm_cb = GenericCallbacks.AtInit(PySDMCallback(
-        "PySDMCallback",
-        solver_config.dg,
-        interpol,
-        mpicomm,
-        pysdm_cw,
-    ))
-
-
-
-    # get aux variables indices for testing
-    q_tot_ind = varsindex(vars_state(model, Auxiliary(), FT), :q_tot)
-    q_vap_ind = varsindex(vars_state(model, Auxiliary(), FT), :q_vap)
-    q_liq_ind = varsindex(vars_state(model, Auxiliary(), FT), :q_liq)
-    q_ice_ind = varsindex(vars_state(model, Auxiliary(), FT), :q_ice)
-    S_liq_ind = varsindex(vars_state(model, Auxiliary(), FT), :S_liq)
-
-    # call solve! function for time-integrator
-    result = ClimateMachine.invoke!(
-        solver_config;
-        diagnostics_config = dgn_config,
-        user_callbacks = (cbvtk, pysdm_cb),
-        check_euclidean_distance = true,
-    )
-
-    # no supersaturation
-    max_S_liq = maximum(abs.(solver_config.dg.state_auxiliary[:, S_liq_ind, :]))
-    @test isequal(max_S_liq, FT(0))
-
-    # qt is conserved
-    max_q_tot = maximum(abs.(solver_config.dg.state_auxiliary[:, q_tot_ind, :]))
-    min_q_tot = minimum(abs.(solver_config.dg.state_auxiliary[:, q_tot_ind, :]))
-    @test isapprox(max_q_tot, qt_0; rtol = 1e-3)
-    @test isapprox(min_q_tot, qt_0; rtol = 1e-3)
-
-    # q_vap + q_liq = q_tot
-    max_water_diff = maximum(abs.(
-        solver_config.dg.state_auxiliary[:, q_tot_ind, :] .-
-        solver_config.dg.state_auxiliary[:, q_vap_ind, :] .-
-        solver_config.dg.state_auxiliary[:, q_liq_ind, :],
-    ))
-    @test isequal(max_water_diff, FT(0))
-
-    # no ice
-    max_q_ice = maximum(abs.(solver_config.dg.state_auxiliary[:, q_ice_ind, :]))
-    @test isequal(max_q_ice, FT(0))
-
-    # q_liq ∈ reference range
-    max_q_liq = maximum(solver_config.dg.state_auxiliary[:, q_liq_ind, :])
-    min_q_liq = minimum(solver_config.dg.state_auxiliary[:, q_liq_ind, :])
-    @test max_q_liq < FT(1e-3)
-    @test isequal(min_q_liq, FT(0))
+    return driver_config, solver_config
 end
-
-main()
